add live_model_base and integrate with slates
