+++ rlclientlib/live_model_slates.cc
@@ -4,64 +4,8 @@
 #include "err_constants.h"
 #include "live_model_impl.h"
 
-#define INIT_CHECK()                                                                                     \
-  do {                                                                                                   \
-    if (!_initialized)                                                                                   \
-    {                                                                                                    \
-      RETURN_ERROR_ARG(nullptr, status, not_initialized, "Library not initialized. Call init() first."); \
-    }                                                                                                    \
-  } while (0);
-
 namespace reinforcement_learning
 {
-live_model_slates::live_model_slates(const utility::configuration& config, error_fn fn, void* err_context,
-    trace_logger_factory_t* trace_factory, data_transport_factory_t* t_factory, model_factory_t* m_factory,
-    sender_factory_t* s_factory, time_provider_factory_t* time_prov_factory)
-{
-  _pimpl = std::unique_ptr<live_model_impl>(
-      new live_model_impl(config, fn, err_context, trace_factory, t_factory, m_factory, s_factory, time_prov_factory));
-}
-
-live_model_slates::live_model_slates(const utility::configuration& config,
-    std::function<void(const api_status&)> error_cb, trace_logger_factory_t* trace_factory,
-    data_transport_factory_t* t_factory, model_factory_t* m_factory, sender_factory_t* s_factory,
-    time_provider_factory_t* time_prov_factory)
-{
-  _pimpl = std::unique_ptr<live_model_impl>(new live_model_impl(
-      config, std::move(error_cb), trace_factory, t_factory, m_factory, s_factory, time_prov_factory));
-}
-
-live_model_slates::live_model_slates(live_model_slates&& other) noexcept
-{
-  std::swap(_pimpl, other._pimpl);
-  _initialized = other._initialized;
-}
-
-live_model_slates::~live_model_slates() = default;
-
-live_model_slates& live_model_slates::operator=(live_model_slates&& other) noexcept
-{
-  std::swap(_pimpl, other._pimpl);
-  _initialized = other._initialized;
-  return *this;
-}
-
-int live_model_slates::init(api_status* status)
-{
-  if (_initialized) { return error_code::success; }
-
-  const auto err_code = _pimpl->init(status);
-  if (err_code == error_code::success) { _initialized = true; }
-
-  return err_code;
-}
-
-std::vector<int> live_model_slates::c_array_to_vector(const int* c_array, size_t array_size)
-{
-  if (c_array == nullptr) { return {}; }
-  return std::vector<int>(c_array, c_array + array_size);
-}
-
 int live_model_slates::request_multi_slot_decision(
     const char* event_id, string_view context_json, unsigned int flags, multi_slot_response& resp, api_status* status)
 {
@@ -142,13 +86,6 @@ int live_model_slates::request_multi_slot_decision(const char* event_id, string_
   return _pimpl->request_multi_slot_decision(event_id, context_json, flags, resp, baseline_vector, status);
 }
 
-// not implemented yet
-int live_model_slates::report_action_taken(const char* event_id, api_status* status)
-{
-  INIT_CHECK();
-  return _pimpl->report_action_taken(event_id, status);
-}
-
 int live_model_slates::report_outcome(const char* event_id, const char* outcome, api_status* status)
 {
   INIT_CHECK();
@@ -160,11 +97,4 @@ int live_model_slates::report_outcome(const char* event_id, float outcome, api_s
   INIT_CHECK();
   return _pimpl->report_outcome(event_id, outcome, status);
 }
-
-int live_model_slates::report_action_taken(const char* primary_id, const char* secondary_id, api_status* status)
-{
-  INIT_CHECK();
-  return _pimpl->report_action_taken(primary_id, secondary_id, status);
-}
-
 }  // namespace reinforcement_learning
